requirements.txt
accelerate==1.11.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
controlnet-aux==0.0.3
cycler==0.12.1
diffusers==0.27.2
easydict==1.13
einops==0.8.1
fastapi==0.108.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gradio==4.21.0
gradio_client==0.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
IOPaint==1.6.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.12.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.11.4
packaging==25.0
pandas==2.3.3
peft==0.7.1
piexif==1.1.3
Pillow==9.5.0
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.7.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
replicate==1.0.7
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
ruff==0.14.5
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.16.3
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.32.0.post1
sympy==1.14.0
tifffile==2025.10.16
timm==1.0.22
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.48.3
typer==0.20.0
typer-config==1.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==11.0.3
wsproto==1.3.1
yacs==0.1.8
zipp==3.23.0
